requirements.txt
Flask
requests
newspaper3k
python-dateutil
